macOS arm64: Stage 4.3 — integer ALU sequences; all integer PPC tests pass
- New sequences: ADD_CARRY, MUL, MUL_HI, DIV, AND_NOT, CNTLZ,
  IS_TRUE/IS_FALSE, MIN (int, signed), SELECT (int), CONTEXT_BARRIER,
  LOAD_LOCAL/STORE_LOCAL (new a64_seq_memory.cc)
- Sub-word (i8/i16) operands are explicitly extended before compares,
  divides, and high-multiplies since upper bits are undefined
- Fix upstream HIR constant-folding bug: unsigned 64-bit MulHi
  sign-extended int64 operands into unsigned __int128 (non-MSVC path
  only; caught by instr_mulhdu _constant cases)
- Emit() logs the unhandled opcode to stderr before asserting (the async
  logger loses the line when abort() fires)

Integer suite: addc/adde/addic/addme/addze/andc/cntlzd/cntlzw/divd/
divdu/divw/divwu/mulhd/mulhdu/mulhw/mulhwu/mulld/mulli/mullw/sld/slw/
srad/sradi/sraw/srawi/srd/srw/subfc/subfe/subfic/subfme/subfze/
seq_branch_carry all 0 failures.

+++ src/xenia/cpu/hir/value.cc
@@ -409,9 +409,13 @@ void Value::MulHi(Value* other, bool is_unsigned) {
       }
 #else
       if (is_unsigned) {
+        // Truncate to uint64_t first: int64_t -> unsigned __int128 sign-
+        // extends negative values across the full 128 bits.
         constant.i64 = static_cast<uint64_t>(
-            (static_cast<unsigned __int128>(constant.i64) *
-             static_cast<unsigned __int128>(other->constant.i64)) >>
+            (static_cast<unsigned __int128>(
+                 static_cast<uint64_t>(constant.i64)) *
+             static_cast<unsigned __int128>(
+                 static_cast<uint64_t>(other->constant.i64))) >>
             64);
       } else {
         constant.i64 = static_cast<uint64_t>(
